+++ packages/app-project/pages/projects/[owner]/[project]/classify/index.js
@@ -1,3 +1,2 @@
-export { default } from '@screens/ClassifyPage'
-export { default as getServerSideProps } from '@helpers/getDefaultPageProps'
+export { default, getServerSideProps } from '@screens/ClassifyPage'
