+++ packages/app-project/pages/projects/[owner]/[project]/classify/workflow/[workflowID]/index.js
@@ -1,10 +1,2 @@
-import getDefaultPageProps from '@helpers/getDefaultPageProps'
+export { default, getServerSideProps } from '@screens/ClassifyPage'
 
-export { default } from '@screens/ClassifyPage'
-
-export async function getServerSideProps({ params, query, req, res }) {
-  const { workflowID } = params
-  const { props: defaultProps } = await getDefaultPageProps({ params, query, req, res })
-  const props = Object.assign({}, { workflowID }, defaultProps)
-  return ({ props })
-}
