+++ packages/app-project/pages/projects/[owner]/[project]/classify/workflow/[workflowID]/subject-set/[subjectSetID]/index.js
@@ -1,10 +1,1 @@
-import getDefaultPageProps from '@helpers/getDefaultPageProps'
-
-export { default } from '@screens/ClassifyPage'
-
-export async function getServerSideProps({ params, query, req, res }) {
-  const { subjectSetID, workflowID } = params
-  const { props: defaultProps } = await defaultServerSideProps({ params, query, req, res })
-  const props = Object.assign({}, { subjectSetID, workflowID }, defaultProps)
-  return ({ props })
-}
+export { default, getServerSideProps } from '@screens/ClassifyPage'
